fix: little improvements
