+++ src/sys/errors.rs
@@ -1,5 +1,29 @@
 use std::fmt;
 
+#[derive(Debug)]
+pub enum Extension {
+    NES,
+    InvalidExtension,
+}
+
+impl Extension {
+    pub fn from_str(ext: &str) -> Self {
+        match ext.to_uppercase().as_str() {
+            "NES" => Extension::NES,
+            _ => Extension::InvalidExtension,
+        }
+    }
+}
+
+impl PartialEq for Extension {
+   fn eq(&self, other: &Self) -> bool {
+        match (self, other) {
+            (Extension::NES, Extension::NES) => true,
+            _ => false,
+        }
+    }
+}
+
 pub struct InvalidROMFile;
 pub struct ErrorOnOpen;
 pub struct ErrorOnRead;
